+++ project/game_components/environment.h
@@ -15,11 +15,11 @@
 struct Environment
 {
     SDL_Texture** textures;
-    size_t        texture_count;
-    size_t*       cur_texture_indexes;
+    unsigned short        texture_count;
+    unsigned short*       cur_texture_indexes;
 
     SDL_FRect*    rects; /// Exists solely for manipulation by `move_component`.
-    size_t        object_count;
+    unsigned short        object_count;
 
     struct Move_Component* move_component;
 };
@@ -27,11 +27,11 @@ struct Environment
 
 /* Predef */
 
-struct Environment        new_environment(char **const texture_paths, const size_t texture_count, const size_t object_count, int *const exit_code);
+struct Environment        new_environment(char **const texture_paths, const unsigned short texture_count, const unsigned short object_count, int *const exit_code);
 void couple_move_component_to_environment(struct Environment *const to, struct Move_Component *const move_component, const struct Vec2 max_offset, int *const exit_code);
 void                     free_environment(struct Environment *const target);
 void        render_environment(const struct Environment *const target);
-void partly_render_environment(const struct Environment *const target, const size_t min_path_pt, size_t max_path_pt);
+void partly_render_environment(const struct Environment *const target, const unsigned short min_path_pt, unsigned short max_path_pt);
 
 
 #endif /// ENVIRONMENT_H
